Ivor the Chronicler
The youngest son of a successful merchant, Ivor was tutored along with his
elder siblings on the road - as the family flitted between trading hubs. Not
gifted with the natural charm and quick minds of his elder siblings, Ivor
quickly lost interest in the business he was never going to inherit. In its
place came a fascination with the history and myths of the continent. Although
his brothers and sisters all learned many tongues - the only language that Ivor
ever took to was Elvish, delighted by the secrets that this ancient people and
their texts could unlock.

As he moved into young adulthood, Ivor announced his unlikely to study with the
famed chronicler monks at Oak-Tree Abbey. A generous donation from his Father
secured him am place, and at 15 left the family to begin life as a novice.
Quiet, occasionally forgetful, Ivor was not quick to make a name for himself -
and when he choose to specialise in the obscure "Cult of the Dragon", he seemed
destined to be no more than a footnote in the venerable institution's famed
Chronicles.

For nearly 20 years Ivor meticulously studied his subject in the Oak-Tree
archives. More often than not, when he did venture out to ancient libraries or
rumoured historical sites he would return empty handed - inevitably generating
gentle mockery from his more successful brothers at the Abbey.