Alternate Universe Ivor: Level_1
The youngest son of a successful merchant, the path of Ivor's life was
altered for ever when his father was convicted of Fraud and embezzlement
after falling foul of a member of the local nobility, Lord Haslow.
Impoverished, and with their widower father gaoled for life, the 3
children suddenly found themselves thrown onto the streets of Baldur's
Gate.

Life was very hard for the first few years, but the young Ivor was sheltered
from the worse of it by his elder siblings who slowly started to crawl onto the
bottom rung of the local underworld. After several years, between Elisa's
prodigious thieving skills and Alexey's Del Boy-esque "talent" for a shady deal,
the family could start to think about surviving the next week rather than until
the next meal.

Both of his elder siblings worried about Ivor's chances of survival. He was more
than able to keep up with Elisa on any thieving mission for the guild, and his
sharp sense and saved them both more than once. None the less he lacked the
confidence and bravado required to flourish in this mirky underworld.

It was Elisa who came up with a solution, coming into some valuable information
she and Alexey were able to blackmail the Abbot of Oak-Tree Abbey to take Ivor
in as a novice. In order to save face, it was agreed that everyone (including)
Ivor would be told that he had entered the monastery on a scholarship.

The Abbot was good to his word, and Ivor was not mistreated. In truth the
timing was perfect - he was after a patsy. The Abbey had started to hear
whispers that the "Cult of the dragon" were stirring into action. Ivor was
given the unglamorous research specialism of the Cult. Given access to the
archives of the monastery and resources to hunt down historical sites Ivor
published prodigiously for nearly 20 years. In truth he discovered little the
leaders of the Order did not already know - but he had become the publish face
of the Order's research into the cult. When the time came, he would be
sacrificed allowing the cult to believe they had removed the one remaining
expert in the field.