Bahamut's Boon of Undectability
The ride back to Waterdeep is a long one, and Ivor falls into a deep
sleep. Wondering through a dream like Neverwinter finds himself in front
of a familiar hooded elder.

"I owe you thanks, without [the party] we would have been powerless
to prevent Tiamat's ascension - for it is only mortals who have the
luxury to defy prophecy."

Ivor glances at the sun, still fixed in endless twilight - the crops of
the Sword case are slowly dying.

"Yes - your work is not done. But the planes can be restored to their
natural order - your bigger concern should be the keeper of history you
broke out of eternal imprisonment."

Ivor's heart sank.

"Still young mortal, my thanks you have. What would you claim as your
reward?".

Ivor paused, he could feel himself getting tired - and not the sort that
could be solved by a good night's sleep. Thinking of Rakshasas that
come back to hunt you again and again after deaths, of terrifying
spiders forged of blades  that hunt from across planes - but mostly of a
world perpetually in need of aid, the answer became clear.

"When all this is over, I wish to retire in peace to a simple life
where none but my true friends may find me."

Bahmut's smiled. So shall it be, that you take the title of Ivor the
Imperceptible.