Intrigue at Mirabar
The party recouperate in Luskan, as preparations are made in the high
tower to send them into the elemental planes. Ivor uses the opportunity
to get onto the property ladder and pays way over the odds for an
admitedly beautiful appartment overlooking the Harbour.

Meanwhile a message arrives for Layton - there's trouble in Mirabar.
There are various factions vying for authority in the Queen's absence.
Things are starting to get our of hand after an entire family were
assassinated overnight - they quite literally went up in flame at a
banquet.

Leaving Hendry and Layton to attempt to disentangle the politicians,
Brigid, Taryn and Ivor slip down to a bar popular with the palace guards
and attempt to get the gossip from the streets.

After a few hours, Ivor gets talking to a female guard who thinks she
might have a bit of a thing for him. The guard leads a small platoon of
guards and was on duty the night of the murders. She saw a giant eagle
fly away from the explosion and attempted to chase it into the City,
where it had appeared to land. However, when they caught up there was no
sign of it at all. She's convinced that rumours of changelings in the
city are true.

As the evening goes on, it becomes apparent that the 3 of them are being
closely watched. The spy later leaves stalking the guard back to her
Barracks. Ivor makes his excuses and slips into the shadows following
the spy - slipping in behind as they dimension door into an abandoned
cathedral. Here the spy drops their illusionary form and a reveals
themselves to be a faceless changeling. As they prepare to dimension
door out again Ivor grabs a dagger from their belt in they hope they can
be scy'd later. Peering in through the temporary portal reveals they
have slipped into some dank underground caves.

Working his way back through the shadows he regroups with Brigid and
they head back to the palace where a feast is being held Layton's
honour. Another changeling has infiltrated the court and fills Layton's
cup with enchanted wine - as soon as it touches her lips it ignites into
a full-blown fireball. The changeling attempts to flee but is pursued by
Ivor who's just arrived with Brigid. Desperate the assassin throws a
dagger at Ivor who simply catches it and takes down the changeling with
her own dagger. The guards arrive and take her away, but almost as soon
as they leave the room the real guards turn up and it is apparent that
there are yet more changelings!

A fight ensues as the party attempt to run them down, and they attempt
to flee into yet another dimension door. Ivor slips into the shadows and
follows - while Layton and Hendry charge through running down the
assassins. Brigid stays to attempt to defuse a bomb left by the fleeing
assassins. Unfortunately this fails and a Huge fireball takes out the
entire wing of the palace, leaving Brigid alive but injured.

In the caves Ivor slips through the shadows to come find the changelings
now pretending to be Dragonborn - manipulating a world weary Aidan
Burner. It appears they have taken a highly creative interpretation of
"ensure the stability" of Mirabar. Throwing his voice Ivor to across
the room, Ivor reveals the deception to Burner. The ancient red dragon
is not fooled by Ivor's attempt at misplacement, but the changelings
flee.

Ivor politely steps from the shadows and acknowledges the old enemy.
Having cut through the remaining assassins Hendry and Layton discover
Ivor sitting and having Tea with the ancient red dragon.